Trying to convert MAXT to double; also deleting A at each iteration
Former-commit-id: 0611b2e2b2cb7514b397443679ccd561146346f0

+++ jComputer.cpp
@@ -7,7 +7,7 @@ using namespace std;
 using namespace arma;
 
 
-JComputer::JComputer(int arg1,int nnOnly, int arg3){
+JComputer::JComputer(double arg1,int nnOnly, int arg3){
 	this->nnOnly = nnOnly;
 }
